IMPLEMENTATION: Change response format to JSON
The service now responds to teh GET request in a JSON format with two fields, namely url and malware.
The url is a string field which is equal to the url passed in teh GET request.
The malware is a boolean field indicating whether the given url is a malware or not.